requirements.txt
pandas==2.2.0
SQLAlchemy==2.0.25
Faker==22.5.1
psycopg2-binary==2.9.9
pytest==8.0.0
pytest-cov==4.1.0
schedule==1.2.1
PyYAML==6.0.1
python-dotenv==1.0.1
